refactor(cli): extract credentials and list runtime bits
