+++ src/lib/nemoclaw-runtime-bridge.ts
@@ -3,25 +3,13 @@
 
 /* v8 ignore start -- transitional bridge until command actions are extracted from src/nemoclaw.ts. */
 
-import type { RecoveryResult } from "./inventory-commands";
-
 export interface SpawnLikeResult {
   status: number | null;
   stdout?: string | Buffer;
   stderr?: string | Buffer;
 }
 
-export interface GatewayRecoveryResult {
-  recovered: boolean;
-}
-
 export interface NemoClawRuntimeBridge {
-  captureOpenshell: (
-    args: string[],
-    opts?: { ignoreError?: boolean; timeout?: number },
-  ) => { status: number | null; output: string };
-  recoverNamedGatewayRuntime: () => Promise<GatewayRecoveryResult>;
-  recoverRegistryEntries: (options?: { requestedSandboxName?: string | null }) => Promise<RecoveryResult>;
   runOpenshell: (
     args: string[],
     opts?: { ignoreError?: boolean; stdio?: import("node:child_process").StdioOptions },
